Added CF
CF is a FORTH clone designed to be used as a scripting engine inside a
terminal. It works exact for the fact that the system api is a bit
small at this point. There is a class included designed to work as a
standalone cf interpreter enhancing the system api by only one
function: a function to exit the interpreter. Also this update include
some changes inside the terminal output system which are makes this
version incompatible to older versions. This version includes a new
build of the library but it isn´t runnable as a standalone jar so you
need to rebuild it if you want to use cf as a standalone scripting
language.