Add ParaDate class
The class represents a datatype designed to store 1 point of time. It
replaces the Date class of standard java because the getters and
setters of java.util.date are marked as @Deprecated (even if there are
no replacing classes for them) which means that they will at some point
disappear.